Move entities into `/users/entities` folder
