+++ src/boards/boards.module.ts
@@ -3,7 +3,7 @@ import { BoardsController } from './boards.controller';
 import { BoardsService } from './boards.service';
 import { Board } from './entities/board.entity';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { User } from '../users/user.entity';
+import { User } from '../users/entities/user.entity';
 import { BoardMapper } from './boards.mapper';
 import { BoardColumnsModule } from 'src/board-columns/board-columns.module';
