+++ src/boards/boards.service.ts
@@ -3,7 +3,7 @@ import { Board } from './entities/board.entity';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Like, Repository } from 'typeorm';
 import { CreateBoardDto } from './dtos/create-board.dto';
-import { User } from '../users/user.entity';
+import { User } from '../users/entities/user.entity';
 import { FindBoardDto } from './dtos/find-board.dto';
 import { UpdateBoardDto } from './dtos/update-board.dto';
 import { BoardColumnsService } from 'src/board-columns/board-columns.service';
