+++ src/users/users.mapper.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@nestjs/common';
-import { User } from './user.entity';
+import { User } from './entities/user.entity';
 import { UserDto } from './dtos/user.dto';
 
 @Injectable()
